--- a/src/libpfm4/docs/man3/libpfm_intel_icx_unc_pcu.3
+++ b/src/libpfm4/docs/man3/libpfm_intel_icx_unc_pcu.3
@@ -20,7 +20,7 @@ Enable edge detection, i.e., count only when there is a state transition from no
 .TP
 .B t
 Set the threshold value. When set to a non-zero value, the counter counts the number
-of IRP clockticks in which the number of occurrences of the event is greater or equal to
+of PCU clockticks in which the number of occurrences of the event is greater or equal to
 the threshold.  This is an integer modifier with values in the range [0:63].
 .TP
 .B i