--- a/src/libpfm4/lib/Makefile
+++ b/src/libpfm4/lib/Makefile
@@ -130,6 +130,7 @@ SRCS   += pfmlib_amd64.c pfmlib_intel_core.c pfmlib_intel_x86.c \
 	  pfmlib_intel_slm.c \
 	  pfmlib_intel_tmt.c \
 	  pfmlib_intel_knl.c \
+	  pfmlib_intel_adl.c \
 	  pfmlib_intel_knl_unc_imc.c \
 	  pfmlib_intel_knl_unc_edc.c \
 	  pfmlib_intel_knl_unc_cha.c \
@@ -294,6 +295,8 @@ INC_X86= pfmlib_intel_x86_priv.h	\
 	events/intel_glm_events.h	\
 	events/intel_icl_events.h	\
 	events/intel_spr_events.h	\
+	events/intel_adl_glc_events.h	\
+	events/intel_adl_grt_events.h	\
 	pfmlib_intel_snbep_unc_priv.h	\
 	events/intel_snbep_unc_cbo_events.h \
 	events/intel_snbep_unc_ha_events.h \