--- a/src/libpfm4/lib/pfmlib_perf_event_pmu.c
+++ b/src/libpfm4/lib/pfmlib_perf_event_pmu.c
@@ -115,7 +115,6 @@ typedef struct {
 #ifndef MAXPATHLEN
 #define MAXPATHLEN	1024
 #endif
-static char debugfs_mnt[MAXPATHLEN];
 
 #define PERF_ATTR_HW 0
 #define PERF_ATTR_SW 0
@@ -127,7 +126,7 @@ static char debugfs_mnt[MAXPATHLEN];
 static perf_event_t *perf_pe = perf_static_events;
 static perf_event_t  *perf_pe_free, *perf_pe_end;
 static perf_umask_t *perf_um, *perf_um_free, *perf_um_end;
-static int perf_pe_count, perf_um_count;
+static int perf_pe_count;
 
 static inline int
 pfm_perf_pmu_supported_plm(void *this)
@@ -171,54 +170,7 @@ perf_attridx2um(int idx, int attr_idx)
 	return um;
 }
 
-/*
- * figure out the mount point of the debugfs filesystem
- *
- * returns -1 if none is found
- */
-static int
-get_debugfs_mnt(void)
-{
-	FILE *fp;
-	char *buffer = NULL;
-	size_t len = 0;
-	char *q, *mnt, *fs;
-	int res = -1;
 
-	fp = fopen("/proc/mounts", "r");
-	if (!fp)
-		return -1;
-
-	while(pfmlib_getl(&buffer, &len, fp) != -1) {
-
-		q = strchr(buffer, ' ');
-		if (!q)
-			continue;
-		mnt = ++q;
-		q = strchr(q, ' ');
-		if (!q)
-			continue;
-		*q = '\0';
-
-		fs = ++q;
-		q = strchr(q, ' ');
-		if (!q)
-			continue;
-		*q = '\0';
-
-		if (!strcmp(fs, "debugfs")) {
-			strncpy(debugfs_mnt, mnt, MAXPATHLEN);
-			debugfs_mnt[MAXPATHLEN-1]= '\0';
-			res = 0;
-			break;
-		}
-	}
-	free(buffer);
-
-	fclose(fp);
-
-	return res;
-}
 
 #define PERF_ALLOC_EVENT_COUNT	(512)
 #define PERF_ALLOC_UMASK_COUNT	(1024)
@@ -298,6 +250,58 @@ retry:
 	perf_pe = new_pe;
 
 	goto retry;
+}
+
+#ifndef CONFIG_PFMLIB_NOTRACEPOINT
+static int perf_um_count;
+static char debugfs_mnt[MAXPATHLEN];
+/*
+ * figure out the mount point of the debugfs filesystem
+ *
+ * returns -1 if none is found
+ */
+static int
+get_debugfs_mnt(void)
+{
+	FILE *fp;
+	char *buffer = NULL;
+	size_t len = 0;
+	char *q, *mnt, *fs;
+	int res = -1;
+
+	fp = fopen("/proc/mounts", "r");
+	if (!fp)
+		return -1;
+
+	while(pfmlib_getl(&buffer, &len, fp) != -1) {
+
+		q = strchr(buffer, ' ');
+		if (!q)
+			continue;
+		mnt = ++q;
+		q = strchr(q, ' ');
+		if (!q)
+			continue;
+		*q = '\0';
+
+		fs = ++q;
+		q = strchr(q, ' ');
+		if (!q)
+			continue;
+		*q = '\0';
+
+		if (!strcmp(fs, "debugfs")) {
+			strncpy(debugfs_mnt, mnt, MAXPATHLEN);
+			debugfs_mnt[MAXPATHLEN-1]= '\0';
+			res = 0;
+			break;
+		}
+	}
+	free(buffer);
+
+	fclose(fp);
+
+	return res;
 }
 
 /*
@@ -536,6 +540,7 @@ gen_tracepoint_table(void)
 	}
 	closedir(dir1);
 }
+#endif /* CONFIG_PFMLIB_NOTRACEPOINT */
 
 static int
 pfm_perf_detect(void *this)
@@ -605,8 +610,10 @@ pfm_perf_init(void *this)
 	 */
 	perf_event_support.pme_count = PME_PERF_EVENT_COUNT;
 
+#ifndef CONFIG_PFMLIB_NOTRACEPOINT
 	/* must dynamically add tracepoints */
 	gen_tracepoint_table();
+#endif
 
 	/* must dynamically add optional hw events */
 	add_optional_events();