--- a/src/libpfm4/perf_examples/task.c
+++ b/src/libpfm4/perf_examples/task.c
@@ -250,12 +250,15 @@ parent(char **arg)
 
 	for(i=0; i < num_fds; i++) {
 		int is_group_leader; /* boolean */
+		uint32_t group_pmu = -1;
 
+		/* we can only group events if the belong to the same PMU */
 		is_group_leader = perf_is_group_leader(fds, i);
 		if (is_group_leader) {
 			/* this is the group leader */
 			group_fd = -1;
-		} else {
+			group_pmu = fds[i].hw.type;
+		} else if (fds[i].hw.type == group_pmu) { /* same PMU */
 			group_fd = fds[fds[i].group_leader].fd;
 		}