… src/libpfm4/lib/events/intel_adl_glc_events.h
… src/libpfm4/lib/events/intel_adl_grt_events.h
… src/libpfm4/tests/validate_x86.c