## Pull Request Description
Update libpfm4. See commit history below. 

NOTE: Support for Intel Alder- and Raptor Lake was added without testing, as the PAPI team currently doesn't have access to these architectures.

Current with
commit f517f1ec8038de00ce8f5fefeeef704e24aa08ae
Author: Stephane Eranian <eranian@gmail.com>
Date:   Thu Feb 29 23:07:21 2024 -0800

    add CONFIG_PFMLIB_NOTRACEPOINT to speedup libpfm4 initialization
    
    When pfmlib_initialize() is run as root and if the debugs is mounted,
    then the library parses all the tracepoints to add them to the perf PMU.
    Depending on the number of tracepoints, this can take a significant amount
    of time even though this may not be needed if no tracepoints is passed.
    In order to speedup pfm_initialize(), the patch adds a compile-time option
    to disable support for tracepoint in the perf PMU.
    
    To deactivate tracepoint support: make CONFIG_PFMLIB_NOTRACEPOINT=y
    
    The default build is unchanged with tracepoints enabled. This patch
    adds an opt-out option.
    
    Signed-off-by: Stephane Eranian <eranian@gmail.com>

commit 816bb547f8997b84f7ef70bb99420f40dc8a984d
Author: Stephane Eranian <eranian@gmail.com>
Date:   Mon Feb 26 22:39:49 2024 -0800

    add Intel Raptorlake PMU support

    Enables support for Raptorlake, Raptorlake P, Raptorlake S.

    Signed-off-by: Stephane Eranian <eranian@gmail.com>

commit 8fa4467ffa7014ed8f1525783b5919b80117beca
Author: Stephane Eranian <eranian@gmail.com>
Date:   Mon Feb 19 22:58:17 2024 -0800

    Add Intel AlderLake Gracemont (E-Core) core PMU support

    Adds core PMU support for Alderlake E-core (gracemont).
    Based on Intel JSON events v1.24 published from
    github.com/intel/perfmon/ADL

    Signed-off-by: Stephane Eranian <eranian@gmail.com>

commit e84a9563f4c93dc6e530dfa55d61b150fbf51510
Author: Stephane Eranian <eranian@gmail.com>
Date:   Wed Feb 14 23:09:49 2024 -0800

    Add Intel AlderLake Goldencove (P-Core) core PMU support

    Adds core PMU support for Alderlake P-core (goldencove).
    Based on Intel JSON events v1.24 published from
    github.com/intel/perfmon/ADL

    Signed-off-by: Stephane Eranian <eranian@gmail.com>

commit 9410619f922facca7dab2406c58fe41a8dd61529
Author: Stephane Eranian <eranian@gmail.com>
Date:   Tue Feb 20 23:22:47 2024 -0800

    update task.c example to handle hybrid

    Cannot group event if they do not belong to the same hardware PMU.

    Signed-off-by: Stephane Eranian <eranian@gmail.com>

commit 2441b263f6f28c0fe80f8cee62cd2e64d75cd433
Author: Stephane Eranian <eranian@gmail.com>
Date:   Wed Feb 28 20:08:30 2024 -0800

    add INTEL_X86_CODE_DUP event flag for Intel PMUs

    To handle the case where two events shared the same code and none
    is because of deprecation. In order to pass validation both events
    with the same code must have that flag set.

    Signed-off-by: Stephane Eranian <eranian@gmail.com>

commit 9669e0d696a98b8b5655186dda8b457113cb0ba2
Author: Stephane Eranian <eranian@gmail.com>
Date:   Mon Feb 19 22:48:44 2024 -0800

    Add support for deprecated events for Intel X86 PMUs

    Adds INTEL_X86_DEPRECATED flag to Intel X86 events.
    Deprecated in this context means there is a newer event
    monitoring the same condition. This is used to mark events
    as deprecated and avoid detecting duplicate event codes.

    Signed-off-by: Stephane Eranian <eranian@gmail.com>

commit b7307408ddb1548271983d1fd7c4f17287d2dc0e
Author: Stephane Eranian <eranian@gmail.com>
Date:   Tue Feb 13 21:58:11 2024 -0800

    fix Intel IcelakeX uncore PCU PMU man page typo

    Was referring to eRP clockticks instead of PCU clockticks.

    Signed-off-by: Stephane Eranian <eranian@gmail.com>

## Author Checklist
- [x] **Description**
_Why_ this PR exists. Reference all relevant information, including _background_, _issues_, _test failures_, etc
- [x] **Commits**
_Commits_ are self contained and only do one thing
_Commits_ have a header of the form: `module: short description`
_Commits_ have a body (whenever relevant) containing a detailed description of the addressed problem and its solution
- [x] **Tests**
The PR needs to pass all the tests