@junyongheo 2024-03-06
Hello, I have access to a Raptor Lake computer and am willing to help test. If anyone can point me in the general direction I can try it out.